move Slider into components/Home
